move the ResponseContent to correct package.
