+++ restlight-jaxrs-provider/src/test/java/io/esastack/restlight/jaxrs/adapter/MessageBodyWriterAdapterTest.java
@@ -23,8 +23,7 @@ import io.esastack.commons.net.http.HttpHeaders;
 import io.esastack.commons.net.http.MediaType;
 import io.esastack.commons.net.netty.http.Http1HeadersImpl;
 import io.esastack.restlight.core.resolver.ResponseEntity;
-import io.esastack.restlight.core.util.Ordered;
-import io.esastack.restlight.server.bootstrap.ResponseContent;
+import io.esastack.restlight.core.resolver.ResponseContent;
 import io.esastack.restlight.server.context.RequestContext;
 import io.esastack.restlight.server.context.impl.RequestContextImpl;
 import io.esastack.restlight.server.core.HttpOutputStream;
